% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{geom_hist_with_image}
\alias{geom_hist_with_image}
\alias{GeomHistWithImage}
\title{Histogram Using Images}
\usage{
geom_hist_with_image(bottomImg = NULL, topImg = NULL, mapping = NULL,
  data = NULL, stat = "identity", position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, bgColor = "white",
  lineColor = 1, lineWidth = 2, lineType = "solid", breaks = "Sturges",
  freq = NULL, ...)
}
\arguments{
\item{bottomImg}{The image "inside" the histogram. This should be an image object
as obtained by \code{png::readPNG}, \code{jpeg::readJPEG},
\code{magick::image_read} etc. You must specify at least this image.}

\item{topImg}{The image "outside" the histogram.}

\item{bgColor}{Background color outside the histogram in case you don't specify
\code{topImg}, defaults to white.}

\item{lineColor}{Line color as accepted by R in the \code{col} parameter, defaults to black.}

\item{lineWidth}{Line width as accepted by R in the \code{lwd} parameter, defaults to 1.}

\item{lineType}{Line type as accepted by R in the \code{lty} parameter, defaults to "solid".}

\item{breaks}{See the \code{breaks} parameter in \code{?hist}, defaults to "Sturges".}

\item{freq}{See the \code{freq} parameter in \code{?hist}.}

\item{...}{Other standard ggplot2 parameters such as mapping, data, stat etc.}
}
\description{
This function allows you to draw a simple histogram of a single variable,
using one or two images
}
\note{
By default \code{geom_hist_with_image} will not add padding in the x and y axes.
This default behavior of ggplot2 does not make sense when it comes to plotting with images.
You can however override this by adding \code{+ xlim(c(0, 100))}. Make sure
the result is meaningful though.
}
\examples{
nyc <- png::readPNG(system.file("extdata", "nyc.png", package = "ggwithimages"))
night_sky <- png::readPNG(system.file("extdata", "night_sky.png", package = "ggwithimages"))
nyc_accidents <- readr::read_csv(system.file("extdata", "nyc_accidents.csv", package = "ggwithimages"))
library(ggplot2)
ggplot(nyc_accidents, aes(n_accidents)) +
  geom_hist_with_image(nyc, night_sky) +
  ylim(c(0, 700))

}
\keyword{datasets}
