% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{geom_line_with_image}
\alias{geom_line_with_image}
\alias{GeomLineWithImage}
\title{Line Plot Using Images}
\usage{
geom_line_with_image(bottomImg = NULL, topImg = NULL, mapping = NULL,
  data = NULL, stat = "identity", position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, bgColor = "white",
  lineColor = 1, lineWidth = 1, lineType = "solid", ...)
}
\arguments{
\item{bottomImg}{The image "below" the line. This should be an image object
as obtained by \code{png::readPNG}, \code{jpeg::readJPEG},
\code{magick::image_read} etc. You must specify at least this image.}

\item{topImg}{The image "above" the line.}

\item{bgColor}{Background color above the line in case you don't specify
\code{topImg}, defaults to white.}

\item{lineColor}{Line color as accepted by R in the \code{col} parameter, defaults to black.}

\item{lineWidth}{Line width as accepted by R in the \code{lwd} parameter, defaults to 1.}

\item{lineType}{Line type as accepted by R in the \code{lty} parameter, defaults to "solid".}

\item{...}{Other standard ggplot2 parameters such as mapping, data, stat etc.}
}
\description{
This function allows you to draw a simple line plot of y vs. x, using one or
two images
}
\note{
By default \code{geom_line_with_image} will not add padding in the x and y axes.
This default behavior of ggplot2 does not make sense when it comes to plotting with images.
You can however override this by adding \code{+ xlim(c(0, 100))}. Make sure
the result is meaningful though.
}
\examples{
annapurna <- png::readPNG(system.file("extdata", "annapurna.png", package = "ggwithimages"))
sky <- png::readPNG(system.file("extdata", "sky.png", package = "ggwithimages"))
kathmandu_hourly_aqi <- readr::read_csv(system.file("extdata", "kathmandu_hourly_aqi.csv", package = "ggwithimages"))
library(ggplot2)
ggplot(kathmandu_hourly_aqi, aes(hour, aqi)) +
  geom_line_with_image(annapurna, sky)

}
\keyword{datasets}
