% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{rotate}
\alias{rotate}
\title{helper function to rotate a 2D matrix 90 degrees clock-wise, n times}
\usage{
rotate(mat, n = 1)
}
\description{
helper function to rotate a 2D matrix 90 degrees clock-wise, n times
}
